pyproject.toml
[project]
name = "potential-flow"
version = "0.1.0"
description = "Joukowski Airfoil Potential Flow Visualizer"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "gunicorn>=25.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-playwright>=0.7.2",
]
